+++ flutternotes/lib/main.dart
@@ -2,6 +2,8 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutternotes/views/login_view.dart';
+import 'package:flutternotes/views/register_view.dart';
+import 'package:flutternotes/views/verify_email_view.dart';
 
 import 'firebase_options.dart';
 
@@ -21,6 +23,10 @@ class MyApp extends StatelessWidget {
         primarySwatch: Colors.blue,
       ),
       home: const Homepage(),
+      routes: {
+        '/login/': (context) => const LoginView(),
+        '/register/': (context) => const RegisterView()
+      },
     );
   }
 }
@@ -30,52 +36,29 @@ class Homepage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('HomePage'),
+    return FutureBuilder(
+      future: Firebase.initializeApp(
+        options: DefaultFirebaseOptions.currentPlatform,
       ),
-      // future builder delays the building of a widget until the future task is accomplished
-      body: FutureBuilder(
-        future: Firebase.initializeApp(
-          options: DefaultFirebaseOptions.currentPlatform,
-        ),
-        //Snsnapshot has the states of the future task
-        builder: (context, snapshot) {
-          switch (snapshot.connectionState) {
-            case ConnectionState.done:
-              final user = FirebaseAuth.instance.currentUser;
-              if (user?.emailVerified == true) {
-                return const VerifyEmailView();
-              } else {
-                return const LoginView();
-              }
-            default:
-              return const Text('Loading..');
-          }
-        },
-      ),
-    );
-  }
-}
-
-class VerifyEmailView extends StatefulWidget {
-  const VerifyEmailView({Key? key}) : super(key: key);
-
-  @override
-  State<VerifyEmailView> createState() => _VerifyEmailViewState();
-}
-
-class _VerifyEmailViewState extends State<VerifyEmailView> {
-  @override
-  Widget build(BuildContext context) {
-    return Column(children: [
-      const Text('Please verify your email address:'),
-      ElevatedButton(
-          onPressed: () async {
+      //Snsnapshot has the states of the future task
+      builder: (context, snapshot) {
+        switch (snapshot.connectionState) {
+          case ConnectionState.done:
             final user = FirebaseAuth.instance.currentUser;
-            await user?.sendEmailVerification();
-          },
-          child: const Text('send email verification'))
-    ]);
+            if (user != null) {
+              if (user.emailVerified == true) {
+                print('verified');
+              } else {
+                return const VerifyEmailView();
+              }
+            } else {
+              return const LoginView();
+            }
+            return const Text('done');
+          default:
+            return const CircularProgressIndicator();
+        }
+      },
+    );
   }
 }
