+++ flutternotes/lib/views/register_view.dart
@@ -30,38 +30,48 @@ class _RegisterViewState extends State<RegisterView> {
 
   @override
   Widget build(BuildContext context) {
-    return Column(children: [
-      TextField(
-        controller: email,
-        keyboardType: TextInputType.emailAddress,
-        decoration: const InputDecoration(hintText: 'Enter your email here'),
-      ),
-      TextField(
-        controller: password,
-        obscureText: true,
-        autocorrect: false,
-        enableSuggestions: false,
-        decoration: const InputDecoration(hintText: 'Enter your password here'),
-      ),
-      ElevatedButton(
-        onPressed: () async {
-          try {
-            final usercredential = await FirebaseAuth.instance
-                .createUserWithEmailAndPassword(
-                    email: email.text, password: password.text);
-            print(usercredential);
-          } on FirebaseAuthException catch (e) {
-            if (e.code == 'weak-password') {
-              print('Weak Password');
-            } else if (e.code == 'email-already-in-use') {
-              print('Email already in use');
-            } else if (e.code == 'invalid-email') {
-              print('Invalid Email');
+    return Scaffold(
+      appBar: AppBar(title: const Text('Register')),
+      body: Column(children: [
+        TextField(
+          controller: email,
+          keyboardType: TextInputType.emailAddress,
+          decoration: const InputDecoration(hintText: 'Enter your email here'),
+        ),
+        TextField(
+          controller: password,
+          obscureText: true,
+          autocorrect: false,
+          enableSuggestions: false,
+          decoration:
+              const InputDecoration(hintText: 'Enter your password here'),
+        ),
+        ElevatedButton(
+          onPressed: () async {
+            try {
+              final usercredential = await FirebaseAuth.instance
+                  .createUserWithEmailAndPassword(
+                      email: email.text, password: password.text);
+              print(usercredential);
+            } on FirebaseAuthException catch (e) {
+              if (e.code == 'weak-password') {
+                print('Weak Password');
+              } else if (e.code == 'email-already-in-use') {
+                print('Email already in use');
+              } else if (e.code == 'invalid-email') {
+                print('Invalid Email');
+              }
             }
-          }
-        },
-        child: const Text('Register'),
-      ),
-    ]);
+          },
+          child: const Text('Register'),
+        ),
+        TextButton(
+            onPressed: () {
+              Navigator.of(context)
+                  .pushNamedAndRemoveUntil('/login/', (route) => false);
+            },
+            child: const Text('Already resgistered? Login here'))
+      ]),
+    );
   }
 }
